Use a separate file for FakeGui
