+++ spec/fake_gui.rb
@@ -1,0 +1,29 @@
+require 'gosu'
+require 'gui'
+
+class FakeGosuWindow
+  def stub(*args)
+    true
+  end
+
+  alias initialize    stub
+  alias draw_quad     stub
+  alias draw_triangle stub
+  alias rotate        stub
+end
+
+GosuWindowBackup = Gosu::Window
+Gosu.send(:remove_const, :Window)
+Gosu::Window = FakeGosuWindow
+
+GuiBackup = Gui
+self.class.send(:remove_const, :Gui)
+
+load 'gui.rb' # reload gui.rb after stubbing
+
+FakeGui = Gui # catch the stubbed gui
+
+self.class.send(:remove_const, :Gui)
+Gui = GuiBackup # restore things
+Gosu.send(:remove_const, :Window)
+Gosu::Window = GosuWindowBackup
