+++ spec/gui_spec.rb
@@ -1,33 +1,4 @@
-require 'gosu'
-require 'gui'
-
-class FakeGosuWindow
-  def stub(*args)
-    true
-  end
-
-  alias initialize    stub
-  alias draw_quad     stub
-  alias draw_triangle stub
-  alias rotate        stub
-end
-
-GosuWindowBackup = Gosu::Window
-Gosu.send(:remove_const, :Window)
-Gosu::Window = FakeGosuWindow
-
-GuiBackup = Gui
-self.class.send(:remove_const, :Gui)
-
-load 'gui.rb' # reload gui.rb after stubbing
-
-FakeGui = Gui # catch the stubbed gui
-
-self.class.send(:remove_const, :Gui)
-Gui = GuiBackup # restore things
-Gosu.send(:remove_const, :Window)
-Gosu::Window = GosuWindowBackup
-
+require 'fake_gui'
 require 'arena'
 require 'ai'
 require 'robot'
